ENGG2600/azure-func/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
numpy==1.19.0
tensorflow==2.2.0
matplotlib==3.4.2
sklearn==0.0
